pyproject.toml
[project]
name = "kinocms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.16.5",
    "asgiref==3.9.1",
    "authlib==1.6.3",
    "azure-common==1.1.28",
    "azure-core==1.35.0",
    "azure-identity==1.24.0",
    "azure-mgmt-core==1.6.0",
    "azure-mgmt-rdbms==10.1.0",
    "azure-mgmt-resource==24.0.0",
    "azure-mgmt-subscription==3.1.1",
    "babel==2.17.0",
    "bcrypt==4.3.0",
    "bidict==0.23.1",
    "blinker==1.9.0",
    "boto3==1.40.26",
    "botocore==1.40.26",
    "brotli==1.1.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cryptography==45.0.7",
    "django==5.2.6",
    "django-bootstrap5==25.2",
    "django-modeltranslation==0.19.17",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "flask==3.1.2",
    "flask-babel==4.0.0",
    "flask-compress==1.18",
    "flask-login==0.6.3",
    "flask-mail==0.10.0",
    "flask-migrate==4.1.0",
    "flask-paranoid==0.3.0",
    "flask-principal==0.4.0",
    "flask-security-too==5.6.2",
    "flask-socketio==5.5.1",
    "flask-sqlalchemy==3.1.1",
    "flask-wtf==1.2.2",
    "google-api-core==2.25.1",
    "google-api-python-client==2.181.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httplib2==0.30.0",
    "idna==3.10",
    "importlib-resources==6.5.2",
    "isodate==0.7.2",
    "itsdangerous==2.2.0",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.3.0",
    "jeepney==0.9.0",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "jsonformatter==0.3.4",
    "keyring==25.6.0",
    "ldap3==2.9.1",
    "libgravatar==1.0.4",
    "libpass==1.9.1.post0",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "more-itertools==10.8.0",
    "msal==1.33.0",
    "msal-extensions==1.3.1",
    "msrest==0.7.1",
    "oauthlib==3.3.1",
    "paramiko==3.5.1",
    "passlib==1.7.4",
    "pillow==11.3.0",
    "proto-plus==1.26.1",
    "protobuf==6.32.0",
    "psutil==7.0.0",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "psycopg2==2.9.10",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pynacl==1.5.0",
    "pyotp==2.9.0",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-engineio==4.12.2",
    "python-slugify==8.0.4",
    "python-socketio==5.13.0",
    "pytz==2025.2",
    "pyzstd==0.17.0",
    "qrcode==8.2",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rich==14.1.0",
    "rsa==4.9.1",
    "s3transfer==0.13.1",
    "secretstorage==3.3.3",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "simple-websocket==1.1.0",
    "six==1.17.0",
    "sqlalchemy==2.0.43",
    "sqlparse==0.5.3",
    "sshtunnel==0.4.0",
    "text-unidecode==1.3",
    "transliterate==1.10.2",
    "typer==0.16.1",
    "typing-extensions==4.15.0",
    "ua-parser==1.0.1",
    "ua-parser-builtins==0.18.0.post1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "user-agents==2.2.0",
    "werkzeug==3.1.3",
    "wsproto==1.2.0",
    "wtforms==3.2.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = [
    "E501",  # Black не может автофиксить строки/комментарии
]
# полезные правила
[tool.ruff.lint.per-file-ignores]
"*/migrations/*.py" = ["E501"]  # миграции Django могут быть длинными
"*/admin.py" = ["N806"]  # формсеты Django используют CamelCase

[tool.ruff.format]
quote-style = "double"